requirements.txt
# Core dependencies
apify-client==1.7.0
openai>=1.0.0
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Utilities
requests==2.31.0
pillow==10.1.0

# Optional: for better performance
# openpyxl==3.1.2  # Para exportar Excel
